+++ axelor-business-project/src/main/java/com/axelor/apps/businessproject/service/PurchaseOrderInvoiceProjectServiceImpl.java
@@ -127,6 +127,7 @@ public class PurchaseOrderInvoiceProjectServiceImpl extends PurchaseOrderInvoice
             public List<InvoiceLine> creates() throws AxelorException {
 
               InvoiceLine invoiceLine = this.createInvoiceLine();
+              invoiceLine.setProject(purchaseOrderLine.getProject());
 
               List<InvoiceLine> invoiceLines = new ArrayList<InvoiceLine>();
               invoiceLines.add(invoiceLine);
@@ -174,6 +175,7 @@ public class PurchaseOrderInvoiceProjectServiceImpl extends PurchaseOrderInvoice
             public List<InvoiceLine> creates() throws AxelorException {
 
               InvoiceLine invoiceLine = this.createInvoiceLine();
+              invoiceLine.setProject(purchaseOrderLine.getProject());
 
               List<InvoiceLine> invoiceLines = new ArrayList<InvoiceLine>();
               invoiceLines.add(invoiceLine);
@@ -202,6 +204,7 @@ public class PurchaseOrderInvoiceProjectServiceImpl extends PurchaseOrderInvoice
             public List<InvoiceLine> creates() throws AxelorException {
 
               InvoiceLine invoiceLine = this.createInvoiceLine();
+              invoiceLine.setProject(purchaseOrderLine.getProject());
 
               List<InvoiceLine> invoiceLines = new ArrayList<InvoiceLine>();
               invoiceLines.add(invoiceLine);